LRNG - add atomic DRNG implementation
The atomic DRNG implementation supports the in-kernel use cases which
request random numbers in atomic contexts. It uses the ChaCha20 DRNG
which is a code base that does not sleep and it provides an interface to
callers that does not sleep. This code will only be compiled when the
LRNG enables the in-kernel drop-in replacement APIs for the existing
random.c code base.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: AKM <73632243+akm-04@users.noreply.github.com>